refactor: move generateName to avoid circular dependency
